moved settings definition and defaults to a new file and setup the display message and log for linting on file change
